Python/Information/Final/requirements.txt
python-dotenv
requests
re
json